+++ lib/flows/calculate-your-child-maintenance.rb
@@ -1,12 +1,43 @@
 status :published
 satisfies_need "100147"
 
-## Q1
+## Q0
+multiple_choice :are_you_paying_or_receiving? do
+  option :pay
+  option :receive
 
+  save_input_as :paying_or_receiving
+
+  calculate :fees_title do
+    PhraseList.new(:"#{paying_or_receiving}_title")
+  end
+
+  calculate :collect_and_pay_service do
+      PhraseList.new(:"#{paying_or_receiving}_collect_and_pay_service")
+    end
+
+  calculate :enforcement_charge do
+    if responses.last == "pay"
+      PhraseList.new(:enforcement_charge)
+    end
+  end
+
+  next_node :how_many_children_paid_for?
+end
+
+## Q1
 multiple_choice :how_many_children_paid_for? do
   option "1_child"
   option "2_children"
   option "3_children"
+
+  precalculate :paying_or_receiving_text do
+    paying_or_receiving == "pay" ? "paying" : "receiving"
+  end
+
+  precalculate :paying_or_receiving_hint do
+    PhraseList.new(:"#{paying_or_receiving}_hint")
+  end
 
   calculate :number_of_children do
     ## to_i will look for the first integer in the string
@@ -22,37 +53,42 @@ multiple_choice :gets_benefits? do
   option "yes"
   option "no"
 
-  calculate :calculator do
-    Calculators::ChildMaintenanceCalculator.new(number_of_children, benefits)
+  precalculate :benefits_title do
+    PhraseList.new(:"#{paying_or_receiving}_benefits")
   end
 
-  next_node do |response|
-    if response == 'yes'
-      :how_many_nights_children_stay_with_payee?
-    else
-      :gross_income_of_payee?
-    end
+  calculate :calculator do
+    Calculators::ChildMaintenanceCalculator.new(number_of_children, benefits, paying_or_receiving)
   end
+
+  next_node_if(:how_many_nights_children_stay_with_payee?, responded_with('yes'))
+  next_node :gross_income_of_payee?
 end
 
 ## Q3
 money_question :gross_income_of_payee? do
-  calculate :flat_rate_amount do
-    calculator.base_amount
+
+  precalculate :income_title do
+    PhraseList.new(:"#{paying_or_receiving}_income")
   end
-  next_node do |response|
+
+  next_node_calculation :rate_type do |response|
     calculator.income = response
-    rate_type = calculator.rate_type
-    if [:nil, :flat].include?(rate_type)
-      "#{rate_type.to_s}_rate_result".to_sym
-    else
-      :how_many_other_children_in_payees_household?
-    end
+    calculator.rate_type
   end
+
+  next_node_if(:nil_rate_result) { rate_type == :nil }
+  next_node_if(:flat_rate_result) { rate_type == :flat }
+  next_node :how_many_other_children_in_payees_household?
 end
 
 ## Q4
 value_question :how_many_other_children_in_payees_household? do
+
+  precalculate :number_of_children_title do
+    PhraseList.new(:"#{paying_or_receiving}_number_of_children")
+  end
+
   calculate :calculator do
     # if converting to int messes things up, raise invalid response
     # this deals with nil input through the API
@@ -71,19 +107,23 @@ multiple_choice :how_many_nights_children_stay_with_payee? do
   option 3
   option 4
 
+  precalculate :how_many_nights_title do
+    PhraseList.new(:"#{paying_or_receiving}_how_many_nights")
+  end
+
   calculate :child_maintenance_payment do
     calculator.number_of_shared_care_nights = responses.last.to_i
     sprintf("%.0f", calculator.calculate_maintenance_payment)
   end
-  next_node do |response|
+
+  next_node_calculation :rate_type do |response|
     calculator.number_of_shared_care_nights = response.to_i
-    rate_type = calculator.rate_type
-    if [:nil, :flat].include?(rate_type)
-      "#{rate_type.to_s}_rate_result".to_sym
-    else
-      :reduced_and_basic_rates_result
-    end
+    calculator.rate_type
   end
+
+  next_node_if(:nil_rate_result) { rate_type == :nil }
+  next_node_if(:flat_rate_result) { rate_type == :flat }
+  next_node :reduced_and_basic_rates_result
 end
 
 outcome :nil_rate_result do
@@ -98,7 +138,16 @@ end
 
 outcome :flat_rate_result do
   precalculate :flat_rate_amount do
-    calculator.base_amount
+    sprintf('%.2f', calculator.base_amount)
+  end
+  precalculate :collect_fees do
+    sprintf('%.2f', calculator.collect_fees)
+  end
+  precalculate :total_fees do
+    sprintf('%.2f', calculator.total_fees(flat_rate_amount, collect_fees))
+  end
+  precalculate :total_yearly_fees do
+    sprintf('%.2f', calculator.total_yearly_fees(collect_fees))
   end
 end
 
@@ -111,4 +160,17 @@ outcome :reduced_and_basic_rates_result do
       rate_type.to_s
     end
   end
+  precalculate :child_maintenance_payment do
+    sprintf('%.2f', child_maintenance_payment)
+  end
+  precalculate :collect_fees do
+    sprintf('%.2f', calculator.collect_fees_cmp(child_maintenance_payment))
+  end
+  precalculate :total_fees do
+    sprintf('%.2f', calculator.total_fees_cmp(child_maintenance_payment, collect_fees))
+  end
+  precalculate :total_yearly_fees do
+    sprintf('%.2f', calculator.total_yearly_fees(collect_fees))
+  end
+
 end
